requirements.txt
transformers
sentence-transformers
faiss-cpu
spacy
nltk